Changed component folder structure
